# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flight Booking API",
    "version": "1.1.0",
    "description": "A production-grade REST API for flight booking and enterprise API automation practice."
  },
  "servers": [
    { "url": "http://localhost:3000/api/v1" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    },
    "schemas": {
...
      "Passenger": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "age": { "type": "integer" },
          "passport": { "type": "string" },
          "seat": { "type": "string" }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "user": { "type": "string" },
          "flight": { "type": "string" },
          "passengers": { "type": "array", "items": { "$ref": "#/components/schemas/Passenger" } },
          "status": { "type": "string", "enum": ["pending", "confirmed", "cancelled", "failed"] },
          "paymentStatus": { "type": "string", "enum": ["pending", "successful", "failed"] }
        }
      },
...
    }
  },
  "security": [
    { "bearerAuth": [] },
    { "ApiKeyAuth": [] }
  ],
  "paths": {
...
    "/bookings": {
      "post": {
        "summary": "Create booking",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Booking" } }
          }
        },
        "responses": { "201": { "description": "Booking created" } }
      },
      "get": {
        "summary": "List all bookings (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of bookings" } }
      }
    },
...
    "/export/bookings": {
      "post": {
        "summary": "Export bookings as CSV (async job)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "202": { "description": "Export job started" } }
      }
    },
...
  }
}